WTEL-1102

+++ app/cc.go
@@ -1,5 +1,5 @@
 package app
 
-func (fm *FlowManager) JoinToInboundQueue(domainId int, callId string, queueId int64, queueName string, priority int) (string, error) {
-	return fm.cc.Member().JoinCallToQueue(int64(domainId), callId, queueId, queueName, priority)
+func (fm *FlowManager) JoinToInboundQueue(domainId int64, callId string, queueId int64, queueName string, priority int) (string, error) {
+	return fm.cc.Member().JoinCallToQueue(domainId, callId, queueId, queueName, priority)
 }
